[CREATE] PushNotificationManagementViewController make Complete
